requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-environ==0.11.2
Pillow==10.2.0
channels==4.0.0
daphne==4.1.0
